IDK why reflection isn't working

+++ eventbus/src/main/kotlin/me/austin/rush/Listener.kt
@@ -13,7 +13,7 @@ interface Listener {
     val target: KClass<*>
 
     /** The priority that the listener will be called upon */
-    val priority: Byte
+    val priority: Int
 
     /**
      * Processes an event passed through this listener.
